refactor(app): remove data-testid from molecules components
remove data-testid from molecules components

closes AUTH-
