+++ app/src/molecules/LiquidDetailCard/LiquidDetailCard.tsx
@@ -215,7 +215,7 @@ export function LiquidDetailCard(props: LiquidDetailCardProps): JSX.Element {
       onClick={handleSelectedValue}
       width="10.3rem"
       minHeight="max-content"
-      data-testid="LiquidDetailCard_box"
+      role="button"
     >
       <Flex
         flexDirection={DIRECTION_COLUMN}
